examples/490-haystack-deepgram-stt-pipeline-python/requirements.txt
deepgram-sdk==6.1.1
haystack-ai==2.27.0
python-dotenv==1.2.2
